+++ app/cmd/cli/node.go
@@ -34,17 +34,19 @@ func init() {
 }
 
 var nodeStakeCmd = &cobra.Command{
-	Use:   "stake <fromAddr> <amount> <RelayChainIDs> <serviceURI> <networkID> <fee>",
-	Short: "Stake a node in the network",
+	Use:   "stake <operatorPublicKey||signerAddress> <amount> <RelayChainIDs> <serviceURI> <outputAddress||signerAddress> <networkID> <fee> <isBefore8.0>",
+	Short: "Stake a node in the network, the signer may be the operator or the output address. The signer must specify the public key of the output or operator",
 	Long: `Stake the node into the network, making it available for service.
 Will prompt the user for the <fromAddr> account passphrase. After the 0.6.X upgrade, if the node is already staked, this transaction acts as an *update* transaction.
 A node can updated relayChainIDs, serviceURI, and raise the stake amount with this transaction.
 If the node is currently staked at X and you submit an update with new stake Y. Only Y-X will be subtracted from an account
-If no changes are desired for the parameter, just enter the current param value just as before`,
-	Args: cobra.ExactArgs(6),
+If no changes are desired for the parameter, just enter the current param value just as before.
+The signer may be the operator or the output address. The signer must specify the public key of the output or operator`,
+	Args: cobra.ExactArgs(8),
 	Run: func(cmd *cobra.Command, args []string) {
 		app.InitConfig(datadir, tmNode, persistentPeers, seeds, remoteCLIURL)
-		fromAddr := args[0]
+		operator := args[0]
+		output := args[4]
 		amount, err := strconv.Atoi(args[1])
 		if err != nil {
 			fmt.Println(err)
@@ -64,13 +66,18 @@ If no changes are desired for the parameter, just enter the current param value 
 		rawChains := reg.ReplaceAllString(args[2], "")
 		chains := strings.Split(rawChains, ",")
 		serviceURI := args[3]
-		fee, err := strconv.Atoi(args[5])
+		fee, err := strconv.Atoi(args[6])
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+		isBefore8, err := strconv.ParseBool(args[7])
 		if err != nil {
 			fmt.Println(err)
 			return
 		}
 		fmt.Println("Enter Passphrase: ")
-		res, err := StakeNode(chains, serviceURI, fromAddr, app.Credentials(pwd), args[4], types.NewInt(int64(amount)), int64(fee), false)
+		res, err := StakeNode(chains, serviceURI, operator, output, app.Credentials(pwd), args[5], types.NewInt(int64(amount)), int64(fee), isBefore8)
 		if err != nil {
 			fmt.Println(err)
 			return
@@ -90,20 +97,25 @@ If no changes are desired for the parameter, just enter the current param value 
 }
 
 var nodeUnstakeCmd = &cobra.Command{
-	Use:   "unstake <fromAddr> <networkID> <fee>",
+	Use:   "unstake <operatorAddr> <fromAddr> <networkID> <fee> <isBefore8.0>",
 	Short: "Unstake a node in the network",
 	Long: `Unstake a node from the network, changing it's status to Unstaking.
 Will prompt the user for the <fromAddr> account passphrase.`,
-	Args: cobra.ExactArgs(3),
+	Args: cobra.ExactArgs(5),
 	Run: func(cmd *cobra.Command, args []string) {
 		app.InitConfig(datadir, tmNode, persistentPeers, seeds, remoteCLIURL)
-		fee, err := strconv.Atoi(args[2])
+		fee, err := strconv.Atoi(args[3])
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+		isBefore8, err := strconv.ParseBool(args[4])
 		if err != nil {
 			fmt.Println(err)
 			return
 		}
 		fmt.Println("Enter Password: ")
-		res, err := UnstakeNode(args[0], app.Credentials(pwd), args[1], int64(fee), false)
+		res, err := UnstakeNode(args[0], args[1], app.Credentials(pwd), args[2], int64(fee), isBefore8)
 		if err != nil {
 			fmt.Println(err)
 			return
@@ -123,11 +135,11 @@ Will prompt the user for the <fromAddr> account passphrase.`,
 }
 
 var nodeUnjailCmd = &cobra.Command{
-	Use:   "unjail <fromAddr> <networkID> <fee>",
+	Use:   "unjail <operatorAddr> <fromAddr> <networkID> <fee> <isBefore8.0>",
 	Short: "Unjails a node in the network",
 	Long: `Unjails a node from the network, allowing it to participate in service and consensus again.
 Will prompt the user for the <fromAddr> account passphrase.`,
-	Args: cobra.ExactArgs(3),
+	Args: cobra.ExactArgs(4),
 	Run: func(cmd *cobra.Command, args []string) {
 		app.InitConfig(datadir, tmNode, persistentPeers, seeds, remoteCLIURL)
 		fee, err := strconv.Atoi(args[2])
@@ -135,8 +147,13 @@ Will prompt the user for the <fromAddr> account passphrase.`,
 			fmt.Println(err)
 			return
 		}
+		isBefore8, err := strconv.ParseBool(args[3])
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
 		fmt.Println("Enter Password: ")
-		res, err := UnjailNode(args[0], app.Credentials(pwd), args[1], int64(fee), false)
+		res, err := UnjailNode(args[0], args[1], app.Credentials(pwd), args[1], int64(fee), isBefore8)
 		if err != nil {
 			fmt.Println(err)
 			return