+++ x/apps/types/msg_test.go
@@ -47,18 +47,18 @@ func TestMsgApp_GetSigners(t *testing.T) {
 	tests := []struct {
 		name string
 		args
-		want sdk.Address
+		want []sdk.Address
 	}{
 		{
 			name: "return signers",
 			args: args{msgAppStake},
-			want: sdk.Address(pk.Address()),
+			want: []sdk.Address{sdk.Address(pk.Address())},
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if got := tt.args.msgAppStake.GetSigner(); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("GetSigner() = %v, want %v", got, tt.want)
+			if got := tt.args.msgAppStake.GetSigners(); !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("GetSigners() = %v, want %v", got, tt.want)
 			}
 		})
 	}
@@ -190,17 +190,17 @@ func TestMsgBeginAppUnstake_GetSigners(t *testing.T) {
 	tests := []struct {
 		name string
 		args
-		want sdk.Address
+		want []sdk.Address
 	}{
 		{
 			name: "return signers",
 			args: args{msgBeginAppUnstake},
-			want: sdk.Address(pk.Address()),
+			want: []sdk.Address{sdk.Address(pk.Address())},
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if got := tt.args.msgBeginAppUnstake.GetSigner(); !reflect.DeepEqual(got, tt.want) {
+			if got := tt.args.msgBeginAppUnstake.GetSigners(); !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("GetSigners() = %v, want %v", got, tt.want)
 			}
 		})
@@ -363,17 +363,17 @@ func TestMsgAppUnjail_GetSigners(t *testing.T) {
 	tests := []struct {
 		name string
 		args
-		want sdk.Address
+		want []sdk.Address
 	}{
 		{
 			name: "return signers",
 			args: args{msgAppUnjail},
-			want: sdk.Address(msgAppUnjail.AppAddr),
+			want: []sdk.Address{sdk.Address(msgAppUnjail.AppAddr)},
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if got := tt.args.msgAppUnjail.GetSigner(); !reflect.DeepEqual(got, tt.want) {
+			if got := tt.args.msgAppUnjail.GetSigners(); !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("GetSigners() = %v, want %v", got, tt.want)
 			}
 		})