+++ x/nodes/keeper/valStateChanges_test.go
@@ -8,6 +8,7 @@ import (
 	abci "github.com/tendermint/tendermint/abci/types"
 	"reflect"
 	"testing"
+	"time"
 )
 
 func TestKeeper_FinishUnstakingValidator(t *testing.T) {
@@ -68,6 +69,9 @@ func TestValidatorStateChange_EditAndValidateStakeValidator(t *testing.T) {
 	updateServiceURL.StakedTokens = stakeAmount
 	updateServiceURL.Chains = newChains
 	updateServiceURL.ServiceURL = "https://newServiceUrl.com"
+	// nil output addresss
+	nilOutputAddress := val
+	nilOutputAddress.OutputAddress = nil
 	//same app no change no fail
 	updateNothingval := val
 	updateNothingval.StakedTokens = stakeAmount
@@ -117,6 +121,14 @@ func TestValidatorStateChange_EditAndValidateStakeValidator(t *testing.T) {
 			err:           types.ErrNotEnoughCoins("pos"),
 		},
 		{
+			name:          "FAIL nil output address",
+			accountAmount: notEnoughCoinsAccount,
+			origApp:       val,
+			amount:        stakeAmount,
+			want:          nilOutputAddress,
+			err:           types.ErrNilOutputAddr("pos"),
+		},
+		{
 			name:          "update nothing for the validator",
 			accountAmount: accountAmount,
 			origApp:       val,
@@ -138,7 +150,7 @@ func TestValidatorStateChange_EditAndValidateStakeValidator(t *testing.T) {
 			if err != nil {
 				t.Fail()
 			}
-			err = keeper.StakeValidator(context, tt.origApp, tt.amount)
+			err = keeper.StakeValidator(context, tt.origApp, tt.amount, tt.origApp.PublicKey)
 			if err != nil {
 				t.Fail()
 			}
@@ -151,7 +163,7 @@ func TestValidatorStateChange_EditAndValidateStakeValidator(t *testing.T) {
 				return
 			}
 			// edit stake
-			_ = keeper.StakeValidator(context, tt.want, tt.want.StakedTokens)
+			_ = keeper.StakeValidator(context, tt.want, tt.want.StakedTokens, tt.want.PublicKey)
 			tt.want.Status = sdk.Staked
 			// see if the changes stuck
 			got, _ := keeper.GetValidator(context, tt.origApp.Address)
@@ -249,7 +261,7 @@ func TestKeeper_StakeValidator(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			k := tt.fields.keeper
-			if got := k.StakeValidator(tt.args.ctx, tt.args.validator, tt.args.amount); !reflect.DeepEqual(got, tt.want) {
+			if got := k.StakeValidator(tt.args.ctx, tt.args.validator, tt.args.amount, tt.args.validator.PublicKey); !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("StakeValidator() = %v, want %v", got, tt.want)
 			}
 		})
@@ -450,3 +462,89 @@ func TestKeeper_WaitToBeginUnstakingValidator(t *testing.T) {
 		})
 	}
 }
+
+func TestKeeper_ValidateUnjailMessage(t *testing.T) {
+	type args struct {
+		ctx sdk.Context
+		k   Keeper
+		v   types.Validator
+		msg types.MsgUnjail
+	}
+	unauthSigner := getRandomValidatorAddress()
+	validator := getStakedValidator()
+	validator.Jailed = true
+	validatorNoOuptut := validator
+	validatorNoOuptut.OutputAddress = nil
+	context, _, keeper := createTestInput(t, true)
+	msgUnjailAuthorizedByValidator := types.MsgUnjail{
+		ValidatorAddr: validator.Address,
+		Signer:        validator.Address,
+	}
+	msgUnjailAuthorizedByOutput := types.MsgUnjail{
+		ValidatorAddr: validator.Address,
+		Signer:        validator.OutputAddress,
+	}
+	msgUnjailUnauthorizedSigner := types.MsgUnjail{
+		ValidatorAddr: validator.Address,
+		Signer:        unauthSigner,
+	}
+	tests := []struct {
+		name string
+		args args
+		want error
+	}{
+		{"Test ValidateUnjailMessage With Output Address & AuthorizedByValidator", args{
+			ctx: context,
+			k:   keeper,
+			v: validator,
+			msg: msgUnjailAuthorizedByValidator,
+		}, nil},
+		{"Test ValidateUnjailMessage With Output Address & AuthorizedByOutput", args{
+			ctx: context,
+			k:   keeper,
+			v: validator,
+			msg: msgUnjailAuthorizedByOutput,
+		}, nil},
+		{"Test ValidateUnjailMessage Without Output Address & AuthorizedByValidator", args{
+			ctx: context,
+			k:   keeper,
+			v: validatorNoOuptut,
+			msg: msgUnjailAuthorizedByValidator,
+		}, nil},
+		{"Test ValidateUnjailMessage Without Output Address & AuthroizedByOutput", args{
+			ctx: context,
+			k:   keeper,
+			v: validatorNoOuptut,
+			msg: msgUnjailAuthorizedByOutput,
+		}, types.ErrUnauthorizedSigner("pos")},
+		{"Test ValidateUnjailMessage Without Output Address & Unauthorized", args{
+			ctx: context,
+			k:   keeper,
+			v: validatorNoOuptut,
+			msg: msgUnjailUnauthorizedSigner,
+		}, types.ErrUnauthorizedSigner("pos")},
+
+		{"Test ValidateUnjailMessage With Output Address & Unauthorized", args{
+			ctx: context,
+			k:   keeper,
+			v: validator,
+			msg: msgUnjailUnauthorizedSigner,
+		}, types.ErrUnauthorizedSigner("pos")},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			keeper.SetValidator(tt.args.ctx, tt.args.v)
+			keeper.SetValidatorSigningInfo(tt.args.ctx, tt.args.v.Address, types.ValidatorSigningInfo{
+				Address:             tt.args.v.Address,
+				StartHeight:         0,
+				Index:               0,
+				JailedUntil:         time.Time{},
+				MissedBlocksCounter: 0,
+				JailedBlocksCounter: 0,
+			})
+			_, err := tt.args.k.ValidateUnjailMessage(tt.args.ctx, tt.args.msg)
+			assert.Equal(t, tt.want, err)
+		})
+	}
+}
+